refactor(engine): extract SlotBarrier into submodule (SPL-38.b)
First mechanical extraction from the SPL-38.a audit. Moves
parallel_apply.rs into parallel_apply/mod.rs and lifts SlotBarrier
into a dedicated slot_barrier.rs sibling, preserving the existing
public surface and DG-2.a/DG-3 future-rename headroom.

No behavioural change: the struct shape, Mutex/Condvar fields, and
method bodies are byte-identical to the prior layout. SlotBarrier
moves to pub(super) visibility (with pub(super) fields) so siblings
in the upcoming SPL-38.c/d/e splits can call into it without
churning the API surface again. The lone test that pokes
barrier.notify directly (flush_workers_survives_spurious_wakeup)
continues to compile unchanged.
